RELATED: RAIL-529 Fix sorting & undo/redo
